rename crypto module to password
